refactor: reorganize components folder
